+++ gdal/frmts/envisat/envisatdataset.cpp
@@ -132,8 +132,8 @@ CPLErr MerisL2FlagBand::IReadBlock( CPL_UNUSED int nBlockXOff,
     }
 
     for( unsigned iImg = 0, iBuf = 0;
-         iImg < nBlockXSize * sizeof(GDT_UInt32);
-         iImg += sizeof(GDT_UInt32), iBuf += nBytePerPixel )
+         iImg < nBlockXSize * (unsigned)sizeof(GDT_UInt32);
+         iImg += (unsigned)sizeof(GDT_UInt32), iBuf += nBytePerPixel )
     {
 #ifdef CPL_LSB
         ((GByte*) pImage)[iImg] = pReadBuf[iBuf + 2];
